lucius: revert previous theme change while still fixing the tabline problem
Using the "CursorLine" highlight style was not a good idea for the status line,
because it conflicts when the cursor line is next to the status line of an
inactive window.

Fortunately, the tabline extension allows for specific color customization.
This explicitly defines a color for a selected tab and a selected buffer.