Add base16_shell theme
base16_shell theme for vim-airline matches an active Base16 Shell
theme.

It should be used in combination with Base16 Shell, which sets the
correct colors in the terminal:
(https://github.com/chriskempson/base16-shell).

The colors used in this theme are within the 21 term colors set by
Base16 Shell.  The theme supports 2 options described in the docs.