base16: do not overwrite StatusLine highlighting
The base16 group contained a reference to the Statusline highlighting
group. That changed highlighting of the group unintentionally and made
the wildmenu unreadable.

Most likely, instead of the 'statusline' group it was meant to use the
airline_c highlighting group. So use this instead.

However there was one additional catch. It used the highlighting from the
vimCommand highlighting group. However that group was only available
when editing vim files, else it would fall back to use the Normal
highlighting group. Since the vimCommand highlighting group however
links to the Statement highlighting group, use that instead. Statement
should always be available.

fixes #158