Remove dracula theme
closes #191 #192

The dracula Vim colorscheme (https://github.com/dracula/vim) comes with
an accompanying airline-theme. However, since the vim-airline-themes
also ships with the dracula theme, this might cause inconsistencies and
cause confusion which colors should be used.

So simply drop the dracula theme from this repository and use the one from
the upstream dracula colorscheme. After all, you need to install the
color scheme and if you do, you already will get the airline-theme with
it.